SDK/requirements.txt
PyYAML
termcolor
jinja2
markdown
watchdog
pytest
